fix(group_pull_search): Find _all_ unregistered copies (#329)
This got away from me a bit, but the goal of this PR was to fix the
group pull_search so that it finds all unregistered copies of a file
when running.

This means:
* in `UpdatableGroup.update_pull`, the short circuit that skipped the
search when a known corrupt file copy was present in the group is
removed.
* in `alpenhorn.io._default_asyncs.group_search_async`, now all nodes in
the group are explicitly checked, instead of stopping the search at the
first match.
* Also, now only unregistered copies found in this way will be marked
for checking (`has_file='M'`) but _all_ such files are found and
checked. (The previous behaviour would find at most one extant file on
disk, and force-check it, regardless of what its state in the database
was).
* Because the async is now directly iterating over the node list, that
created a need for a standard way for the GroupIO layer to present the
list of nodes it has in the group. As a result I've implemented an
I/O-layer `nodes` property on GroupIO, and changed the old `set_nodes`
method into the setter for this property. The list provided by the
`nodes` property is sorted in priority order (which may be different for
different I/O classes), with earlier nodes more likely to have I/O
performed on them. Under the hood, individial I/O classes are still
allowed to store the node data however they like.
* Finally, the introduction of the `nodes` property means I've been able
to get rid of the private `_nodes` property in `UpdateableGroup` where
it was cacheing the node list (because it was otherwise hard to retrieve
that after setting it).

Though this ended up being more involved that I had hoped (I originally
set out to just delete two lines to accomplish the first point), I think
this smooths out some rough edges when it comes to node management in
groups, as well as fixing the problem I originally found (corrupt files
hiding unregistered files).