refactor: Move io and scheduler modules under daemon
When I first subdivided the code, there was still the idea that
the client was going to support doing I/O, which meant it was
going to need access to both the I/O framework and the scheduler.

But since the CLI refactor removed all I/O from it, that's no logner
the case.  So, let's move the `scheduler` and `io` modules under
`daemon`, since that's the only part that needs them now.

(The only part of this change I don't like is all the top-level
`alpenhorn` submodules now only begin either with a "c" or a "d",
which seems ... odd.)