refactor: Move scheduler under daemon (#335)
When I first subdivided the code, there was still the idea that the
client was going to support doing I/O, which meant it was going to need
to instantiate its own task scheduler.

But since the CLI refactor removed all I/O from it, that's no logner the
case. So, let's move the `scheduler` and module under `daemon`, since
that's the only part that needs it.

(This is just the `scheduler` part of #332. I'm going to deal with `io`
differently than I did there.)