refactor(io): Split `filesize` method into two.
The `actual=False` part (which never returns None) is kept in `filesize`
and the `actual=True` part (which _may_ return None) is now
`storage_used`.

In addition to the use I'll make of this in the next commit, I think
this also removes confusion over what "actual size" means.  To wit:
the `filesize` value is stored in `ArchiveFile.size_b`; the
`storage_used` value is stored in `ArchiveFileCopy.size_b`.