refactor(copy_request_done): Replase `io` with `node_to` and `size`.
This replaces the `io` parameter in `copy_request_done` with two
parameters: `node_to`, the StorageNode handling the copy request,
and `size`, which is either an actual size or a callable which can
compute the size.  In general, this results in the call:
```
  copy_request_done(..., io=io_instance, ...)
```
being changed into
```
  copy_request_done(..., node_to=io_instance.node,
                         size=io_instance.storage_used, ...)
```
but there is now the potential for the io_instance to simply provide the
"storage_used" size, instead of a callable.

The goal here is to remove explicit dependence on the I/O framework from
this file.