feat(schema_version): Implement rich comparisons
This retools the `schema_version` function in `alpenhorn.db` to
permit specifying "rich" comparisons for component schema versions
along the lines of "<3" or ">=2" or even, ">1,<4".  (All schema versions
are integers).

This functionality is provided for third-party extensions which may
need to decide some things about the database schema they have access
to.