refactor: Move DB connection into `start_alpenhorn`
The goal here is to move more of the "standard" alpenhorn
start-up code to a single place.

Changes to the `start_alpenhorn` parameters:
* `cli` now defaults to `True`
* a new bool `check_schema`, which also defaults to True, permits
  disabling the schema version check.

The immediate motivation here is more prep-work for the Exension
API, but the changes to the parameters is more to support an interactive
mode for using alpenhorn, which I'm trying to develop in the `doctest`
branch.

i.e. to do direct queries of the data index, you should be able to use
the alpenhorn code interactively with something like:
```
>>> from alpenhorn.common.util import start_alpenhorn
>>> from alpenhorn.db import ArchiveAcq
>>> start_alpenhorn()
>>> ArchiveAcq.select().where(name="acq_to_find").execute()
```
(where `start_alpenhorn` sweeps under the rug of all the necessary
third-party set up to get alpenhorn working).

There's no change to the operation of the daemon here, but in the CLI,
the database connection is moved up to the top-level `click` group.
The effect is you can't do anything with the CLI anymore if database
connection fails.  Previously you could at least run, say, `alpenhorn
acq --help`.

This does leave a somewhat awkward schema check in the CLI command
group, which I'm not super keen with, but it'll do unless I figure out
a slicker way to allow `alpenhorn db init` to skip the schema check.