feat(schema_version): Implement rich comparisons (#339)
This retools the `schema_version` function in `alpenhorn.db` to permit
specifying "rich" comparisons for component schema versions along the
lines of "<3" or ">=2" or even, ">1,<4". (All schema versions are
integers).

This functionality is provided for third-party extensions which may need
to decide some things about the database schema they have access to.

The only notable change from the previous behaviour is: previously if
you asked to check the schema version for a missing component, the check
would always fail, regardless of the value you specified in
`component_version`. Now, the schema version for such a missing
component is simply treated as zero, and comparison proceeds as normal.