refactor(io): Provide I/O classes directory, rather than dynamic import
It used to be that when you specified an I/O class like "SomeClass" in
the database, that alpenhorn would:

a) lower-case that and try to import either "alpenhorn.io.someclass" or
   a extension module with the name "someclass"
b) once we have the module, then try to access the symbol "SomeClassNodeIO"
  or "SomeClassGroupIO" (depending on whether this is a node or a
  group), which is hopefully a class name, inside that module.

This worked, and was simple to implement when using `alpenhorn.io`
modules, but it did mean we had to be weirdly specific about naming
conventions when third-parties created I/O extensions.

Here is a reversal of this preference: instead of forcing third-party
I/O extensions to conform to the `alpenhorn.io` way of doing things, now
`alpenhorn.io` is changed to provide something looking more like an I/O
Extension.

Each I/O class implemented by `alpenhorn.io` now provides access to the
node and/or group I/O class they provide via a `namedtuple` called
`InternalIO`, which provides a pared-down version of what an I/O
Extension looks like (or will look like, once the Extension API is
finished).

A list of all such internal I/O pseudo-extensions is provided by
`alpenhorn.io` and this list is used to initialise the list of I/O
extensions kept by the `extload` module.

This change vastly simplifies the creation of I/O instances in the
daemon since now there's just a dict of I/O classes keyed by the value
of the Database `io_class` column.  This removes most of the
restrictions on what constitutes a valid I/O class name.