test: Create a full-fledged Fake-FS watchdog Observer
Up till now, we've been able to get away with just using the
PollingObserver directly to monitor the fake filesystem, because it
wasn't being initialised before the fake filesystem was created.
But the previous change breaks this lucky happenstance, meaning the
PollingObserver can no longer be used.

The problem is the PollingEmitter, which is used by the PollingObserver
to generate filesytem events.  The definition of the PollingEmitter
binds both `os.stat` and `os.scandir` in its keyword parameter defaults
(which happens at import time).

So now in the `mock_observer` fixture, we re-implemenet the
`PollingEmitter` in the most trivial way so that we can rebind these
keyword defaults to the PyFakeFS version of `os`.

This implementation requires `watchdog >= 4`.