feat: Extension API (#344)
This change creates the public Extension API in `alpenhorn.extensions`
which is used by third-parties to create Alpenhorn extensions.

Extension _modules_ haven't changed much: they're still listed in the
"extensions" section of the alpenhorn config and still must provide a
`register_extensions` function hook. However, instead of the dict that
this hook used to return, now it just returns a list of `Extension`
instances.

Instantiating classes makes it easier to be flexible with different
kinds of extensions, while collecting the common extension code in one
place.

Effectively, I suppose, this changes what is meant by the term
"Alpenhorn Extension". Previously, an extension was a _module_ which
provided certain functionality (described by the dict returned by
`register_extension`). Now extensions are class instances providing that
functionality, and the contents of the dict are now passed to the class
initialisers. To reflect this (and to prevent old-style extension
modules from getting loaded), the extension hook has been pluralised to
`register_extensions`.

Operationally, this change has meant extension loading has been broken
up in to two parts: `find_extensions`, responsible for loading the
listed extension modules and collecting the Extension instances they
provide, and then an `init_extensions` (run in several parts) for
actually activating the extensions in alpenhorn.

---------

Co-authored-by: Liam Gray <liam.j.gray@usask.ca>