fix(queue): Use deque instead of set for FIFO-key lists
The `FairMultiFIFOQueue` (FMFQ) was "fair" in the sense that FIFOs
with fewer in-progress tasks are prioritised over FIFOs with
more in-progress tasks.  This makes priority-balancing between
nodes (=FIFOs) work in the multi-worker case, since idle workers
will preferentially be given tasks for under-served nodes.

However, this breaks down in the single- (or no-)worker case, which
is how we're running on SciNet.  In this case, because there's only
one consumer of the queue, whenever a `get` is performed, _all_
FIFOs are in the zero-in-progress-tasks bin.

Previously these bins were `set`s, which provided no inherent priority,
but in practice seem to favour one particular FIFO (I'd guess it's
ordering them by hash or something).  The effect is one FIFO would
monopolise the single worker.

So, this changes the FMFQ to make it "fair" even in the single consumer
case: now the in-progress count bins are also deques (FIFOs).  FIFO
keys are added to the tail of the deque and searching for a job to
give to a worker happens from the head, meaning FIFOs which have
recently been active are deprioritised, which makes a big difference
on SciNet (with only one worker) but almost no change on fir (or wind)
where multiple workers are used (since the primary prioritisation is
still number of in-progress tasks).